+++ src/main/java/ua/beauty/calendar/domain/Procedure.java
@@ -17,7 +17,8 @@ import javax.persistence.Id;
 public class Procedure {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    // private Long id;
+    private Integer id;
 
     @NonNull
     private String name;
